DefaultClfApp/requirements.txt
streamlit==1.24.1
pandas==1.5.3
matplotlib==3.7.1
numpy>=1.21
google_api_python_client==1.12.8
google-cloud-aiplatform==1.28.1


